Data and stacks are now relocated in MMU
Turned out I was smoking crack when deciding on a mem map, so there's a
new section in the PDE (called "Section 0") which is sufficient to map
all of the basic stuff like kernel code, stacks and PDEs.

Removed all the eBL() stuff because B and BL are always relative anyway,
duh.

Moved fns out of arm.h because they mess up -O0 (because they'll end up
out-of-line at the top of piboot.c which ruins the entry point)