Renamed extraKernelSources to sources, reworked hosted build
Entailed adding new config.fullyHosted option which assumes everything
should be built against the system headers (will be used for building luac)

Added a config.lua option in addition to klua and ulua. Just means include
the lua sources, but not klua.c or ulua.c. Is set automatically if klua or
ulua is defined in config