Plumbed our statically-linked modules into Lua's 'require'
Modules all get a separate env, nothing gets to write into the global env,
and require("x") when run within a module will set _ENV.x to the contents
of the module. This is slightly stricter than standard Lua which sticks
everything in the global env unless you tell it otherwise (and doesn't set
_ENV.x or even _G.x).

This is all implemented in the new uklua.c, which is used by both klua and
ulua. test.lua has been updated to assert all the tables are setup the
way they're supposed to be.