Pid and process name now tracked and exposed user-side.
Unlike how I originally intended, we do this via BSS since we have to
support that anyway for malloc, and it means the linker does more of the
work for us. Also it avoids page-colouring restrictions. Added a fn
mmu_mapKernelPageInProcess() even though we now don't actually need it.

Since we have to map at least a page for BSS, and malloc uses nowhere near
that amount, the extra copy of process name etc user-side is effectively
free because there's nothing else we can use that memory for.