Added some support for thread scheduling.
Currently, we still have one thread. However, this thread can now
block on waiting for a char from the uart while interrupts are enabled.

In other words, we now support enabling interrupts in SVC mode for the
purpose of implementing thread yielding. When they are ready to run again,
they get returned straight back to user mode. Which is quite nice because
it means we don't have to save the SVC stack when switching out a thread.
And having interrupts enabled while a thread is blocked is really important
for implementing preemptive multitasking.

Also added some improved diagnostics to the exception mode handlers and
moved them to boot.c as they aren't Pi-specific.