Corrected TTBCR to reflect what was actually implemented:
We only allow 1 page for the PT for the user PTs, meaning a process's user
PTs are limited to 1MB, which is 256 pages, meaning max user address space
is 256 MB (1 page PT per MB). In order to restore TTBCR=2 we'd need 4 pages
per user process which would involve some slight rejigging of the memory
map which I don't want to do right now.

Also corrected KUserStacksBase which was plain pointing to the wrong place.