Cleaned up module native initialisation code.
Instead of a hard-coded check for the module name in uklua.c loaderFn(),
modules specify whether they have native init code in build.lua - a pointer
to this function is part of the autogenerated code. The function must be
named init_module_MODULENAME and must be a lua_CFunction. This allows
loaderFn to call it through Lua meaning it doesn't need to know about
specific init functions.

Also simplified some metatable code.