Improved thread_requestComplete() semantics slightly
It no longer nulls request->thread, only request->userPtr. This is useful
to keep track of the thread a request is (or was) associated with.

Also split out the writing the result from the thread signalling - so that
things that handle the result themselves can call thread_requestSignal()
directly.

Also improved switch_process() API to be a bit easier to use.