Added concept of shared pages to kernel.
These will be used by the IPC mechanism (coming soon to a commit near you!)

This commit also adds the data structures for Server objects because they
are closely connected to shared pages. However it doesn't include the rest
of the server code, which is included in the IPC commit.