Changed how SVC stack is implemented.
Previously there was a single supervisor stack shared among all threads
when they were making an SVC call. This prevented even the possibility of
supporting preemption of threads in supervisor mode.

Now, each thread has its own separate 4KB stack for use in SVC mode, and
the first task of the SVC handler is to set r13 appropriately. The
dedicated kernel stack (used only on boot and maybe in the future for fast
execs) is now also only 4KB.

As part of this change, user registers are now pushed to the *user* stack
when performing an exec, rather than to the kernel stack. This means that
we don't store user registers in Thread->savedRegisters when in the
EBlockedFromSvc state. All of these changes are intended to make it easier
to support preemptible SVCs in the future.