Interrupts are now enabled during SVC calls
This has required a fair number of changes. With a view to minimising the
amount of time when interrupts have to be disabled, there are now DFCs to
perform tasks to long to do inline in an IRQ handler.

A number of places have switched to using the atomic APIs to handle IRQs
interrupting SVC code. Uses of thread_requestComplete in IRQs replaced with
dfc_requestComplete.

Thread exit now handled using a DFC as well, so there is an extra thread
state (EDying) to handle the time between the thread exiting and it being
cleaned up.