Fixed inconsistency in how large compiled vs runtime constants were handled.
Boiled down to the difference in versions of strtol being used between
luac and lupi. luac used the standard system strtol which as per the
spec correctly clamps 0x80000000 and above to 0x7FFFFFFF on a 32-bit
system. This is per the spec but is undesirable in the context of lua
which doesn't have the concept of an unsigned type so on our 32-bit
lua_Number=int system would mean any numbers above INT_MAX wouldn't be
representable as positive hex constants. This would make things like the
symbols module very messy as we'd have to output addresses as their
equivalent negative signed integer format.

In the interests of having the lua runtime accept 32-bit unsigned hex
constants (even if subsequent comparison operators consider them as signed
which brings its own set of problems), we use our implementation of strtol
in luac. This ignores the clamping requirement and means that at least the
behaviour of runtime and compile-time constants is the same.

This is made possible by splitting strtol into a separate file strtol.c and
including that file in the luac build, which will override the library
definition.

It also means that advanced constant syntax accepted by C90 strtol like
0x123.456 which previously would have been accepted in a constant if a
module was compiled, but not otherwise and not if parsed from a string
variable, is now disallowed across the board.