Added a switch_process to dfc_threadExit
Now that I look, I'm not totally convinced it's either (a) needed, or (b)
even possible to reach this code with currentProcess not already correct.
But never mind, this is what I've been running with and there's no harm to
it since when the DFC finishes the DFC thread always calls reschedule()
which will switch processes as needed.