Cleaned up implementation of ASSERT() macro
It now leaves less gumf visible in the crash debugger, and puts less code
inline at the site of the ASSERT() call, meaning it's easier to debug.

Now you'll never see an assertion-failure-related function appear in the
crashRegisters, nor will the stack (from sp to top) ever contain any of the
data used in the implementation of the assertion failure handler.