Cleaned up --listing option to use -save-temps
This means that line numbers from errors in the assembly stage can actually
be correlated with the original source because we save the actual file
that the assembler used, rather than creating another one in a separate
invokation and hoping it matches.

It also simplifies the build code, which is nice.