Fixed PendSV's failure to save and restore registers.
Also fixed incorrect saveCurrentRegistersForThread() call when a thread's
timeslice expires (which never happens currently because it's disabled, but
this bug is I think why it was disabled).

Simplified sysTick handler.
Also fixed some potential race conditions in svc blocking.
Also commented out the rescheduleNeededOnSvcExit checks in pendSV because
of some weird timing issues that I haven't got to the bottom of. Short
version is it works with this code removed...