Fixed typo in WFI() macro which meant it has NEVER worked...
This means the Pi has been spinning in a tight loop in reschedule() rather
than sleeping until the next interrupt. Oops...

Also fixed ulua.c's panicFn to not call hang() because that's a kernel-only
fn, replaced with an exec_waitForAnyRequest() loop. And fixed
exec_driverConnect() to return KErrNotFound rather than asserting if asked
to load a non-existant driver.